Before
```
go test ./core/crypto/vrf/p256 -bench=.
goos: darwin
goarch: amd64
pkg: github.com/google/keytransparency/core/crypto/vrf/p256
BenchmarkEvaluate/1_goroutines-12         	     300	   5494696 ns/op
BenchmarkEvaluate/2_goroutines-12         	     500	   2918234 ns/op
BenchmarkEvaluate/4_goroutines-12         	    1000	   1615380 ns/op
BenchmarkEvaluate/8_goroutines-12         	    1000	   1266080 ns/op
BenchmarkEvaluate/16_goroutines-12        	    1000	   1400027 ns/op
BenchmarkEvaluate/32_goroutines-12        	    1000	   1469442 ns/op
BenchmarkEvaluate/64_goroutines-12        	    2000	   1512431 ns/op
BenchmarkEvaluate/128_goroutines-12       	   10000	   1693480 ns/op
PASS
ok  	github.com/google/keytransparency/core/crypto/vrf/p256	35.473s
```

After
```
go test ./core/crypto/vrf/p256 -bench=.
goos: darwin
goarch: amd64
pkg: github.com/google/keytransparency/core/crypto/vrf/p256
BenchmarkEvaluate/1_goroutines-12         	   10000	    189376 ns/op
BenchmarkEvaluate/2_goroutines-12         	   20000	     95095 ns/op
BenchmarkEvaluate/4_goroutines-12         	   30000	     48795 ns/op
BenchmarkEvaluate/8_goroutines-12         	   50000	     32782 ns/op
BenchmarkEvaluate/16_goroutines-12        	   50000	     33574 ns/op
BenchmarkEvaluate/32_goroutines-12        	   50000	     36663 ns/op
BenchmarkEvaluate/64_goroutines-12        	   50000	     36138 ns/op
BenchmarkEvaluate/128_goroutines-12       	   50000	     36754 ns/op
PASS
ok  	github.com/google/keytransparency/core/crypto/vrf/p256	18.597s
```